GMT_DOS-CLIENTS_FEM force_extension module & get/set panics
 - put all HP Force and extentension interface implementation
into force_extension module
 - now Get::get and Set::set will panic if the input or
the output is missing
